pdf_tts/requirements.txt
requests
PyPDF2
python-dotenv